Refactor python code from bin to pylib
